More warning removal
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@999 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/rick/warnings/kernel/platform/unix/FileSystem.cpp
@@ -75,7 +75,7 @@
 extern char achRexxCurDir[ CCHMAXPATH+2 ];    /* Save current working direct */
 
 const char *SysFileExtension(const char *);
-RexxString * LocateProgram(RexxString *, const char *[], LONG);
+RexxString * LocateProgram(RexxString *, const char *[], int);
 const char * SearchFileName(const char *, char);
 FILE * SysBinaryFilemode(FILE *);
